requirements.txt
amqp==5.3.1 ; python_version >= "3.11"
asgiref==3.9.1 ; python_version >= "3.11"
asttokens==3.0.0 ; python_version >= "3.11"
async-timeout==5.0.1 ; python_version >= "3.11" and python_full_version < "3.11.3"
attrs==25.3.0 ; python_version >= "3.11"
billiard==4.2.1 ; python_version >= "3.11"
celery==5.5.3 ; python_version >= "3.11"
cfgv==3.4.0 ; python_version >= "3.11"
click-didyoumean==0.3.1 ; python_version >= "3.11"
click-plugins==1.1.1.2 ; python_version >= "3.11"
click-repl==0.3.0 ; python_version >= "3.11"
click==8.2.1 ; python_version >= "3.11"
colorama==0.4.6 ; (sys_platform == "win32" or platform_system == "Windows") and python_version >= "3.11"
coverage==7.10.3 ; python_version >= "3.11"
decorator==5.2.1 ; python_version >= "3.11"
distlib==0.4.0 ; python_version >= "3.11"
django-cors-headers==4.7.0 ; python_version >= "3.11"
django-extensions==4.1 ; python_version >= "3.11"
django-filter==25.1 ; python_version >= "3.11"
django-redis==6.0.0 ; python_version >= "3.11"
django-silk==5.4.0 ; python_version >= "3.11"
django==5.2.5 ; python_version >= "3.11"
djangorestframework==3.16.1 ; python_version >= "3.11"
drf-spectacular==0.28.0 ; python_version >= "3.11"
drf-standardized-errors==0.15.0 ; python_version >= "3.11"
executing==2.2.0 ; python_version >= "3.11"
factory-boy==3.3.3 ; python_version >= "3.11"
faker==37.5.3 ; python_version >= "3.11"
filelock==3.18.0 ; python_version >= "3.11"
gprof2dot==2025.4.14 ; python_version >= "3.11"
gunicorn==23.0.0 ; python_version >= "3.11"
identify==2.6.13 ; python_version >= "3.11"
inflection==0.5.1 ; python_version >= "3.11"
iniconfig==2.1.0 ; python_version >= "3.11"
ipython-pygments-lexers==1.1.1 ; python_version >= "3.11"
ipython==9.4.0 ; python_version >= "3.11"
jedi==0.19.2 ; python_version >= "3.11"
jsonschema-specifications==2025.4.1 ; python_version >= "3.11"
jsonschema==4.25.0 ; python_version >= "3.11"
kombu==5.5.4 ; python_version >= "3.11"
matplotlib-inline==0.1.7 ; python_version >= "3.11"
nodeenv==1.9.1 ; python_version >= "3.11"
packaging==25.0 ; python_version >= "3.11"
parso==0.8.4 ; python_version >= "3.11"
pexpect==4.9.0 ; python_version >= "3.11" and sys_platform != "win32" and sys_platform != "emscripten"
platformdirs==4.3.8 ; python_version >= "3.11"
pluggy==1.6.0 ; python_version >= "3.11"
pre-commit==4.3.0 ; python_version >= "3.11"
prompt-toolkit==3.0.51 ; python_version >= "3.11"
psycopg2-binary==2.9.10 ; python_version >= "3.11"
ptyprocess==0.7.0 ; python_version >= "3.11" and sys_platform != "win32" and sys_platform != "emscripten"
pure-eval==0.2.3 ; python_version >= "3.11"
pygments==2.19.2 ; python_version >= "3.11"
pytest-cov==6.2.1 ; python_version >= "3.11"
pytest-django==4.11.1 ; python_version >= "3.11"
pytest-factoryboy==2.8.1 ; python_version >= "3.11"
pytest-mock==3.14.1 ; python_version >= "3.11"
pytest==8.4.1 ; python_version >= "3.11"
python-dateutil==2.9.0.post0 ; python_version >= "3.11"
python-dotenv==1.1.1 ; python_version >= "3.11"
pyyaml==6.0.2 ; python_version >= "3.11"
redis==6.4.0 ; python_version >= "3.11"
referencing==0.36.2 ; python_version >= "3.11"
rpds-py==0.27.0 ; python_version >= "3.11"
ruff==0.12.8 ; python_version >= "3.11"
setuptools==80.9.0 ; python_version >= "3.11"
six==1.17.0 ; python_version >= "3.11"
sqlparse==0.5.3 ; python_version >= "3.11"
stack-data==0.6.3 ; python_version >= "3.11"
traitlets==5.14.3 ; python_version >= "3.11"
typing-extensions==4.14.1 ; python_version >= "3.11"
tzdata==2025.2 ; python_version >= "3.11"
uritemplate==4.2.0 ; python_version >= "3.11"
vine==5.1.0 ; python_version >= "3.11"
virtualenv==20.33.1 ; python_version >= "3.11"
wcwidth==0.2.13 ; python_version >= "3.11"
